fix(favorites) : #MAG-350 fix view by board displaying boards without favorite magnets (+ refacto'd)
