+++ frontend/src/components/magnets-collection/MagnetsCollectionModal.scss
@@ -90,3 +90,7 @@ ul.grid {
 .copy-icon {
   margin-bottom: 0.25em;
 }
+
+ul {
+  list-style-type: none;
+}
